Split `MetadataCmd` into its own file
